refactor logos out to own function
